Re-enable Git commit signing
I've disabled this in the past as setting up the GPG toolchain was a
nightmare on different environments and machines, and I struggled to
make it work with devcontainers which are a key (ha, ha) part of my
workflow these days. Now that Git supports signing using SSH keys and
Github can verify these, the friction is low enough for the benefits
of commit verification to be worth it for me again.